+++ frontend/src/pages/account/settings/_mock.ts
@@ -1,22 +1,7 @@
 // eslint-disable-next-line import/no-extraneous-dependencies
 import type {Request, Response} from 'express';
 
-const city = require('./geographic/city.json');
-const province = require('./geographic/province.json');
-
-function getProvince(_: Request, res: Response) {
-  return res.json({
-    data: province,
-  });
-}
-
-function getCity(req: Request, res: Response) {
-  return res.json({
-    data: city[req.params.province],
-  });
-}
-
-function getCurrentUse(req: Request, res: Response) {
+function getCurrentUse(_req: Request, res: Response) {
   return res.json({
     data: {
       name: 'Serati Ma',
@@ -55,16 +40,6 @@ function getCurrentUse(req: Request, res: Response) {
       notifyCount: 12,
       unreadCount: 11,
       country: 'China',
-      geographic: {
-        province: {
-          label: '浙江省',
-          key: '330000',
-        },
-        city: {
-          label: '杭州市',
-          key: '330100',
-        },
-      },
       address: '西湖区工专路 77 号',
       phone: '0752-268888888',
     },
@@ -75,6 +50,4 @@ function getCurrentUse(req: Request, res: Response) {
 export default {
   // 支持值为 Object 和 Array
   'GET  /api/accountSettingCurrentUser': getCurrentUse,
-  'GET  /api/geographic/province': getProvince,
-  'GET  /api/geographic/city/:province': getCity,
 };